initial jest testing
