+++ frontend/app/package.json
@@ -17,7 +17,8 @@
     "lint:fix": "npm run lint -- --fix",
     "prettier": "npx prettier --config .prettierrc.json -w ./src",
     "prettier:check": "npx prettier --config .prettierrc.json --check ./src",
-    "prepare": "cd ../.. && husky install frontend/app/.husky"
+    "prepare": "cd ../.. && husky install frontend/app/.husky",
+    "test-jest": "NODE_ENV=test yarn jest --coverage --no-cache --transformIgnorePatterns \"./frontend/app/svg/\""
   },
   "resolutions": {
     "react-error-overlay": "6.0.9",
@@ -195,8 +196,39 @@
     "typescript": "^5.0.3"
   },
   "jest": {
-    "transformIgnorePatterns": [
-      "node_modules/(?!@ngrx|(?!deck.gl)|ng-dynamic)"
-    ]
+    "collectCoverage": true,
+    "collectCoverageFrom": [
+      "./src/components/**/*.js",
+      "./src/helpers/**/*.ts",
+      "./src/leaderboard/**/*.ts",
+      "./src/pages/**/*.ts",
+      "./src/people/**/*.ts",
+      "./src/tribes/**/*.ts"
+    ],
+    "coverageThreshold": {
+      "global": {
+        "lines": 0
+      },
+      "./src/people/utils/": {
+        "lines": 0
+      },
+      "./src/helpers/": {
+        "lines": 66
+      }
+    },
+    "moduleFileExtensions": [
+      "ts",
+      "js",
+      "json",
+      "node"
+    ],
+    "transform": {
+      "^.+\\.tsx?$": "ts-jest"
+    },
+		"globals": {
+    "window": { "location": {"host": "localhost"}},
+"testEnvironment": "jsdom"
+
+  }
   }
 }
